requirements.txt
scrapy
scrapy-playwright
python-dotenv
# playwright install